patch-1.12 Async Programming update
I have added an example of Async programming in the form of a mad jester who will tell you a random joke fetched from a API using async and await. I have dramatically changed the front end design with a much more 'black' design as it highlights the ? items on the map.